+++ animation_nodes/nodes/matrix/c_utils.pyx
@@ -137,6 +137,45 @@ cdef void scaleFromMatrix(Vector3 *scale, Matrix4 *m):
         scale.y *= -1.0
         scale.z *= -1.0
 
+def extractMatricesXBasis(Matrix4x4List matrices):
+    cdef Vector3DList bases = Vector3DList(length = len(matrices))
+    cdef Matrix4 *_matrices = matrices.data
+    cdef Vector3 *_bases = bases.data
+    cdef Py_ssize_t i
+
+    for i in range(len(bases)):
+        _bases[i].x = _matrices[i].a11
+        _bases[i].y = _matrices[i].a21
+        _bases[i].z = _matrices[i].a31
+
+    return bases
+
+def extractMatricesYBasis(Matrix4x4List matrices):
+    cdef Vector3DList bases = Vector3DList(length = len(matrices))
+    cdef Matrix4 *_matrices = matrices.data
+    cdef Vector3 *_bases = bases.data
+    cdef Py_ssize_t i
+
+    for i in range(len(bases)):
+        _bases[i].x = _matrices[i].a12
+        _bases[i].y = _matrices[i].a22
+        _bases[i].z = _matrices[i].a32
+
+    return bases
+
+def extractMatricesZBasis(Matrix4x4List matrices):
+    cdef Vector3DList bases = Vector3DList(length = len(matrices))
+    cdef Matrix4 *_matrices = matrices.data
+    cdef Vector3 *_bases = bases.data
+    cdef Py_ssize_t i
+
+    for i in range(len(bases)):
+        _bases[i].x = _matrices[i].a13
+        _bases[i].y = _matrices[i].a23
+        _bases[i].z = _matrices[i].a33
+
+    return bases
+
 # Replicate Matrix
 ###############################################